Update app.py
Modified the app.py Gradio UI as follows:

Can enable cpu_offload and quantize_8bit from the web UI (they are enabled by default) - I did this because originally the app.py file did NOT support these flags and it was taking me 90 MINUTES !!! to generate a single image on my 4090. Now it only takes 30 seconds! :)

UI loads faster as it doesn't load the model until the user clicks on Generate.

Automatically saves the outputs to the /results folder.

Shows the seed of the last generation for the user to be able to copy it into the seed for generation box.